+++ CHANGELOG.md
@@ -6,11 +6,11 @@
 * Permissions for Expense class filter (find-po-line-plugin). Refs UIOR-678.
 * Implement Keyboard shortcuts modal. Refs UINV-257.
 * Compile Translation Files into AST Format. Refs UINV-251.
+* Allow user to link an invoice line to a PO line after the invoice line has been created. Refs UINV-193.
 
 ## [2.3.2]https://github.com/folio-org/ui-invoice/tree/v2.3.2) (2021-06-02)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v2.3.1...v2.3.2)
 * Invoice is only approved when user clicks 'approve and pay'. Refs UINV-262.
-* Allow user to link an invoice line to a PO line after the invoice line has been created. Refs UINV-193.
 
 ## [2.3.1]https://github.com/folio-org/ui-invoice/tree/v2.3.1) (2021-04-14)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v2.3.0...v2.3.1)